Remove warnings.
git-svn-id: https://fuse-emulator.svn.sourceforge.net/svnroot/fuse-emulator/trunk@2177 6c7d1b9a-d430-0410-8293-a2b96dacb39b

+++ libspectrum/szx.c
@@ -263,8 +263,8 @@ read_cfrp_chunk( libspectrum_snap *snap, libspectrum_word version GCC_UNUSED,
 
   if( page > 63 ) {
     libspectrum_print_error( LIBSPECTRUM_ERROR_CORRUPT,
-			     "%s:read_ramp_chunk: unknown page number %d",
-			     __FILE__, page );
+			     "%s:read_ramp_chunk: unknown page number %lu",
+			     __FILE__, (unsigned long)page );
     free( data );
     return LIBSPECTRUM_ERROR_CORRUPT;
   }
@@ -288,8 +288,8 @@ read_ramp_chunk( libspectrum_snap *snap, libspectrum_word version GCC_UNUSED,
 
   if( page > 15 ) {
     libspectrum_print_error( LIBSPECTRUM_ERROR_CORRUPT,
-			     "%s:read_ramp_chunk: unknown page number %d",
-			     __FILE__, page );
+			     "%s:read_ramp_chunk: unknown page number %lu",
+			     __FILE__, (unsigned long)page );
     free( data );
     return LIBSPECTRUM_ERROR_CORRUPT;
   }
